[DAD-387] refactor: 스크랩 생성 API 응답, 요청 Dto 파일명 변경
